Convert CreditCard::VirtualCardEnrollmentType enum to enum class
Bug: 1418432
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4703512
Auto-Submit: Prakhar Shrivastav <p.shri@samsung.com>
Commit-Queue: Prakhar Shrivastav <p.shri@samsung.com>
Cr-Commit-Position: refs/heads/main@{#1172990}

+++ components/autofill/core/browser/webdata/autofill_table.cc
@@ -3750,7 +3750,8 @@ void AutofillTable::AddMaskedCreditCards(
     masked_insert.BindString(index++, card.issuer_id());
     masked_insert.BindInt64(index++, card.instrument_id());
     masked_insert.BindInt(index++, card.virtual_card_enrollment_state());
-    masked_insert.BindInt(index++, card.virtual_card_enrollment_type());
+    masked_insert.BindInt(index++, static_cast<int>(
+                                   card.virtual_card_enrollment_type()));
     masked_insert.BindString(index++, card.card_art_url().spec());
     masked_insert.BindString16(index++, card.product_description());
     masked_insert.Run();
